CDPCP-1601. set-password now uses credentials from control plane
The setPassword call now uses the password hash and keys from ums
when possible. Note that we plan to deprecate / remove this endpoint
soon, but for now the UI is still calling it and this behavior
will keep the UMS keys consistent with the keys in FreeIPA, which
is desirable for user keytab generation.