CDPCP-1533. Add operations to usersyncstatus
First part to the rolling update of the freeipa database schema
for tracking user sync status.

This commit adds the requested and successful operations to the
usersync status. The new lastrequestedoperation field is preferred
but lastsyncstarttime field is still used as a
fallback when calculating whether a stack is "cool" for the purposes
of automatic user sync polling.